refactor: Move image nodes to individual files
Implements ML-597
